+++ packages/terra-action-footer/src/terra-dev-site/doc/example/centered/SingleAction.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import Hyperlink from 'terra-hyperlink';
 // eslint-disable-next-line import/no-extraneous-dependencies, import/no-unresolved, import/extensions
 import CenteredActionFooter from 'terra-action-footer/lib/CenteredActionFooter';
 import ExampleTemplate from '../../common/ExampleTemplate';
@@ -6,7 +7,9 @@ import ExampleTemplate from '../../common/ExampleTemplate';
 export default () => (
   <ExampleTemplate>
     <CenteredActionFooter
-      center={<a href="/#/site/action-footer/centered">A link</a>}
+      center={<Hyperlink href="#">A link</Hyperlink>}
     />
   </ExampleTemplate>
 );
+
+//      center={<a href="/#/site/action-footer/centered">A link</a>}